chore(fv): Refactor file structure
Refactored file structure. Moved almost all functions in different files
so the overall file structure could seem more coherent.
Left some functions in the mod.rs file because they seem to be used in
almost every other file.

All tests pass.
