+++ compiler/noirc_evaluator/src/ssa/verus_vir_gen/expr_to_vir/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod exprs;
+pub(crate) mod params;
+pub(crate) mod patterns;
+pub(crate) mod types;
