+++ tooling/nargo_cli/src/cli/fv_cmd.rs
@@ -1,6 +1,5 @@
 use std::{
     io::Write,
-    ops::Not,
     process::{Command, Stdio},
 };
